requirements.txt
pytest == 3.6.3
selenium == 3.13.0
allure-pytest == 2.4.1